feat: Improve macro parsing and output
Quoted arguments are now automatically determined again, and the output
will be quoted when a comma is found in an argument.

This does not match fully with Logseq, which will display quoted
arguments as-is in the macro templates. However unescaping and
normalizing arguments makes it easier to work with macros without having
to think about escaping as the consumer.